Address review feedback and fix CI test failures
